fix(@angular-ru/ng-excel-builder): refactor and use @angular-ru/common as dependency
