+++ packages/ng-excel-builder/package.json
@@ -27,6 +27,7 @@
         "postversion": "npe version \"$(npe version)\" --package=\"$PWD/dist/library/package.json\""
     },
     "peerDependencies": {
+        "@angular-ru/common": ">=13.19.0",
         "@angular/common": ">=8.0.0 <11.0.0",
         "@angular/core": ">=8.0.0 <11.0.0"
     },
