+++ packages/ng-excel-builder/src/public-api.ts
@@ -1,7 +1,5 @@
-export {
-    ExcelBuilderModule,
-    ExcelBuilderService,
-    WebWorkerThreadService,
-    ExcelWorkbook,
-    ExcelWorksheet
-} from './excel';
+export { ExcelBuilderService } from './excel-builder.service';
+export { ExcelBuilderModule } from './excel-builder.module';
+export { ExcelWorksheet } from './symbols';
+export { ExcelWorkbook } from './symbols';
+export { WebWorkerThreadService } from './web-worker-thread.service';
